requirements.txt
openai
python-dotenv
requests
TTS
pvporcupine
pyaudio
